Refactor asset expander
 * Move AssetBundle classes into same file as AssetExpander
 * Remove RailsAssetBundle superclass
 * Remove arguments that are always passed the same value
 * Prefer extending a module to instance_eval
 * Prefer flat_map to map{}.flatten
