+++ lib/jasmine.rb
@@ -8,7 +8,6 @@ jasmine_files = ['base',
                  'command_line_tool',
                  'page',
                  'path_mapper',
-                 'asset_bundle',
                  'asset_pipeline_mapper',
                  'asset_expander',
                  'result',
